moving avatar menu into its own component
